mv _wait_and_warmup
